Ndryshime te Komenti dhe Ankesa endpoints!

+++ backend/micro-services/ankesa/src/main/java/com/localnews/ankesa/Services/AnkesaService.java
@@ -1,5 +1,8 @@
 package com.localnews.ankesa.Services;
 
+import com.localnews.ankesa.Client.UserClient;
+import com.localnews.ankesa.DTO.AnkesaRequest;
+import com.localnews.ankesa.DTO.UserDTO;
 import com.localnews.ankesa.Entity.Ankesa;
 import com.localnews.ankesa.Repositories.AnkesaRepository;
 import jakarta.ws.rs.NotFoundException;
@@ -10,6 +13,7 @@ import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
+import javax.naming.NoPermissionException;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,12 +22,34 @@ public class AnkesaService {
 
     @Autowired
     private AnkesaRepository ankesaRepository;
+    @Autowired
+    private UserClient userClient;
 
+    public Ankesa saveAnkesa(Ankesa ankesa, String token) {
 
+        UserDTO user = userClient.getUserByJwt(token);
 
-    public Ankesa saveAnkesa(Ankesa ankesa) {
+        ankesa.setUserId(user.getId());
 
         return ankesaRepository.save(ankesa);
+    }
+
+    @CacheEvict(value = "ankesa", key="#id")
+    public Ankesa updateAnkesa(Integer id, AnkesaRequest request, String token) throws NoPermissionException {
+        Ankesa updated = ankesaRepository.findById(id).orElseThrow(
+                ()-> new NotFoundException("Ankesa nuk u gjete!")
+        );
+
+        UserDTO user = userClient.getUserByJwt(token);
+
+        if (updated.getUserId() != user.getId())
+            throw new NoPermissionException("Nuk keni autorizim per te bere update ankesen!");
+
+        updated.setTema(request.getTema());
+        updated.setLloji(request.getLloji());
+        updated.setTeksti(request.getTeksti());
+
+        return ankesaRepository.save(updated);
     }
 
     public List<Ankesa> getAllAnkesa() {
@@ -38,7 +64,16 @@ public class AnkesaService {
         );
     }
     @CacheEvict(value = "ankesa", key="#id")
-    public void deleteAnkesa(Integer id) {
+    public void deleteAnkesa(Integer id, String token) throws NoPermissionException {
+        UserDTO user = userClient.getUserByJwt(token);
+
+        Ankesa ankesa = ankesaRepository.findById(id).orElseThrow(
+                () -> new NotFoundException("Ankesa nuk u gjete!")
+        );
+
+        if (user.getId() != ankesa.getUserId() && !user.getRole().toString().equals("ADMIN"))
+            throw new NoPermissionException("Nuk keni autorizim per te fshire kete koment!");
+
         ankesaRepository.deleteById(id);
     }
     @CacheEvict(value = "ankesa", key="#id")
